filter pending cases - more refactoring
