--- src/main/java/org/broadinstitute/consent/http/service/DatabaseElectionCaseAPI.java
@@ -1,179 +1,0 @@
-package org.broadinstitute.consent.http.service;
-
-import com.mongodb.BasicDBObject;
-import com.mongodb.client.FindIterable;
-import org.apache.commons.collections.CollectionUtils;
-import org.broadinstitute.consent.http.db.ConsentDAO;
-import org.broadinstitute.consent.http.db.DACUserDAO;
-import org.broadinstitute.consent.http.db.DataSetDAO;
-import org.broadinstitute.consent.http.db.ElectionDAO;
-import org.broadinstitute.consent.http.db.UserRoleDAO;
-import org.broadinstitute.consent.http.db.VoteDAO;
-import org.broadinstitute.consent.http.db.mongo.MongoConsentDB;
-import org.broadinstitute.consent.http.enumeration.ElectionStatus;
-import org.broadinstitute.consent.http.enumeration.ElectionType;
-import org.broadinstitute.consent.http.enumeration.VoteStatus;
-import org.broadinstitute.consent.http.enumeration.VoteType;
-import org.broadinstitute.consent.http.models.DataSet;
-import org.broadinstitute.consent.http.models.Election;
-import org.broadinstitute.consent.http.models.PendingCase;
-import org.broadinstitute.consent.http.models.Vote;
-import org.broadinstitute.consent.http.models.dto.DataOwnerCase;
-import org.broadinstitute.consent.http.util.DarConstants;
-import org.bson.Document;
-import org.bson.types.ObjectId;
-
-import javax.ws.rs.NotFoundException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-public class DatabaseElectionCaseAPI extends AbstractPendingCaseAPI {
-
-    private ElectionDAO electionDAO;
-    private VoteDAO voteDAO;
-    private DACUserDAO userDAO;
-    private ConsentDAO consentDAO;
-    private DataSetDAO dataSetDAO;
-    private final MongoConsentDB mongo;
-
-    public static void initInstance(ElectionDAO electionDAO, VoteDAO voteDAO, DACUserDAO userDAO, UserRoleDAO userRoleDAO, ConsentDAO consentDAO , MongoConsentDB mongoDB, DataSetDAO dataSetDAO) {
-        PendingCaseAPIHolder.setInstance(new DatabaseElectionCaseAPI(electionDAO, voteDAO, userDAO, userRoleDAO, consentDAO, mongoDB, dataSetDAO));
-
-    }
-
-    private DatabaseElectionCaseAPI(ElectionDAO electionDAO, VoteDAO voteDAO, DACUserDAO userDAO, UserRoleDAO userRoleDAO, ConsentDAO consentDAO, MongoConsentDB mongoDB, DataSetDAO dataSetDAO) {
-        this.electionDAO = electionDAO;
-        this.voteDAO = voteDAO;
-        this.userDAO = userDAO;
-        this.consentDAO = consentDAO;
-        this.mongo = mongoDB;
-        this.dataSetDAO = dataSetDAO;
-    }
-
-    @Override
-    public List<PendingCase> describeDataRequestPendingCases(Integer dacUserId) throws NotFoundException {
-        List<Election> elections = isChairPerson(dacUserId)  ?  electionDAO.findElectionsByTypeAndFinalAccessVoteChairPerson(ElectionType.DATA_ACCESS.getValue(),false) : electionDAO.findElectionsWithFinalVoteByTypeAndStatus(ElectionType.DATA_ACCESS.getValue(), ElectionStatus.OPEN.getValue());
-        List<PendingCase> pendingCases = new ArrayList<>();
-        if (elections != null) {
-            for (Election election : elections) {
-                Vote accessVote = voteDAO.findVoteByElectionIdAndDACUserId(election.getElectionId(),
-                        dacUserId);
-                if (accessVote == null) {
-                    continue;
-                }
-                Integer rpElectionId = electionDAO.findRPElectionByElectionAccessId(election.getElectionId());
-                PendingCase pendingCase = new PendingCase();
-                Boolean isReminderSent;
-                if(Objects.nonNull(rpElectionId)) {
-                    Election rpElection = electionDAO.findElectionWithFinalVoteById(rpElectionId);
-                    Vote rpVote = voteDAO.findVoteByElectionIdAndDACUserId(rpElectionId, dacUserId);
-                    isReminderSent = (accessVote.getIsReminderSent() || rpVote.getIsReminderSent()) ? true : false;
-                    pendingCase.setRpElectionId(rpElectionId);
-                    pendingCase.setAlreadyVoted(accessVote.getVote() != null && rpVote.getVote() != null);
-                    pendingCase.setElectionStatus(rpElection.getStatus().equals(ElectionStatus.FINAL.getValue()) && election.getStatus().equals(ElectionStatus.FINAL.getValue()) ? ElectionStatus.FINAL.getValue() : ElectionStatus.OPEN.getValue());                 // if it's already voted, we should collect vote or do the final election vote
-                    pendingCase.setRpVoteId(rpVote.getVoteId());
-                    pendingCase.setStatus(accessVote.getVote() == null || rpVote.getVote() == null ? VoteStatus.PENDING.getValue() : VoteStatus.EDITABLE.getValue());
-
-                }else{
-                    isReminderSent = (accessVote.getIsReminderSent());
-                    pendingCase.setAlreadyVoted(accessVote.getVote() != null);
-                    pendingCase.setElectionStatus(election.getStatus().equals(ElectionStatus.FINAL.getValue()) ? ElectionStatus.FINAL.getValue() : ElectionStatus.OPEN.getValue());                 // if it's already voted, we should collect vote or do the final election vote
-                    pendingCase.setStatus(accessVote.getVote() == null ? VoteStatus.PENDING.getValue() : VoteStatus.EDITABLE.getValue());
-                }
-                setGeneralFields(pendingCase, election, accessVote, isReminderSent);
-                setFinalVote(dacUserId, election, pendingCase);
-                pendingCases.add(pendingCase);
-            }
-        }
-        return pendingCases;
-    }
-
-    @Override
-    public List<DataOwnerCase> describeDataOwnerPendingCases(Integer dataOwnerId) {
-        List<Election> elections = electionDAO.getElectionByTypeAndStatus(ElectionType.DATA_SET.getValue(), ElectionStatus.OPEN.getValue());
-        List<DataOwnerCase> pendingCases = new ArrayList<>();
-        if (elections != null) {
-            for (Election election : elections) {
-                DataOwnerCase dataOwnerCase = new DataOwnerCase();
-                List<Vote> dataOwnerVotes = voteDAO.findVotesByElectionIdAndType(election.getElectionId(), dataOwnerId, VoteType.DATA_OWNER.getValue());
-                if(CollectionUtils.isNotEmpty(dataOwnerVotes)){
-                    dataOwnerVotes.stream().forEach(v -> {
-                        BasicDBObject query = new BasicDBObject().append(DarConstants.ID, new ObjectId(election.getReferenceId()));
-                        FindIterable<Document> dataAccessRequest = mongo.getDataAccessRequestCollection().find(query);
-                        DataSet dataSet = dataSetDAO.findDataSetById(election.getDataSetId());
-                        dataOwnerCase.setAlias(dataSet.getAlias());
-                        dataOwnerCase.setDarCode(dataAccessRequest != null ?  dataAccessRequest.first().get(DarConstants.DAR_CODE).toString() : null);
-                        dataOwnerCase.setDataSetId(dataSet.getDataSetId());
-                        dataOwnerCase.setDataSetName(dataSet.getName());
-                        dataOwnerCase.setVoteId(v.getVoteId());
-                        dataOwnerCase.setAlreadyVoted(v.getVote() != null);
-                        dataOwnerCase.setReferenceId(election.getReferenceId());
-                        dataOwnerCase.setHasConcerns(v.getHasConcerns());
-                        pendingCases.add(dataOwnerCase);
-                    });
-                }
-            }
-        }
-        return pendingCases;
-    }
-
-
-    private void setFinalVote(Integer dacUserId, Election election, PendingCase pendingCase) {
-        if (pendingCase.getAlreadyVoted()) {
-            Vote chairPersonVote = voteDAO.findChairPersonVoteByElectionIdAndDACUserId(
-                    election.getElectionId(), dacUserId);
-            if (chairPersonVote != null) {
-                pendingCase.setIsFinalVote(chairPersonVote.getVote() != null);
-            }
-        } else {
-            pendingCase.setIsFinalVote(false);
-        }
-    }
-
-    private PendingCase setGeneralFields(PendingCase pendingCase , Election election, Vote vote, boolean isReminderSent) {
-        List<Vote> votes = voteDAO.findDACVotesByElectionId(election.getElectionId());
-        List<Vote> pendingVotes = voteDAO.findPendingVotesByElectionId(election.getElectionId());
-        pendingCase.setTotalVotes(votes.size());
-        pendingCase.setVotesLogged(votes.size() - pendingVotes.size());
-        pendingCase.setReferenceId(election.getReferenceId());
-        if (election.getElectionType().equals(ElectionType.DATA_ACCESS.getValue())) {
-            BasicDBObject query = new BasicDBObject().append(DarConstants.ID, new ObjectId(election.getReferenceId()));
-            FindIterable<Document> dataAccessRequest = mongo.getDataAccessRequestCollection().find(query);
-            if(dataAccessRequest.first() != null) {
-                pendingCase.setFrontEndId(dataAccessRequest.first().get(DarConstants.DAR_CODE).toString());
-                pendingCase.setProjectTitle(dataAccessRequest.first().get(DarConstants.PROJECT_TITLE).toString());
-            }
-        }else{
-             pendingCase.setFrontEndId(consentDAO.findConsentById(election.getReferenceId()).getName());
-             pendingCase.setConsentGroupName(consentDAO.findConsentById(election.getReferenceId()).getGroupName());
-        }
-        pendingCase.setLogged(setLogged(pendingCase.getTotalVotes(), pendingCase.getVotesLogged()));
-        pendingCase.setAlreadyVoted(pendingCase.getAlreadyVoted() == null ? vote.getVote() != null : pendingCase.getAlreadyVoted());
-        pendingCase.setStatus(vote.getVote() == null ? VoteStatus.PENDING.getValue() : VoteStatus.EDITABLE.getValue());
-        pendingCase.setVoteId(vote.getVoteId());
-        pendingCase.setIsReminderSent(isReminderSent);
-        pendingCase.setCreateDate(election.getCreateDate());
-        pendingCase.setElectionStatus(pendingCase.getElectionStatus() == null ? election.getStatus() : pendingCase.getElectionStatus());
-        pendingCase.setElectionId(election.getElectionId());
-        return pendingCase;
-    }
-
-
-    private String setLogged(Integer totalVotes, Integer loggedVotes) {
-        StringBuilder logged = new StringBuilder();
-        logged.append(loggedVotes)
-                .append("/")
-                .append(totalVotes);
-        return logged.toString();
-    }
-
-    private boolean isChairPerson(Integer dacUserId) {
-        boolean isCherperson = false;
-        if (userDAO.checkChairpersonUser(dacUserId) != null) {
-            isCherperson = true;
-        }
-        return isCherperson;
-    }
-
-}
