--- src/main/java/org/broadinstitute/consent/http/service/PendingCaseAPI.java
@@ -1,14 +1,0 @@
-package org.broadinstitute.consent.http.service;
-
-import org.broadinstitute.consent.http.models.PendingCase;
-import org.broadinstitute.consent.http.models.dto.DataOwnerCase;
-
-import javax.ws.rs.NotFoundException;
-import java.util.List;
-
-public interface PendingCaseAPI {
-
-    List<PendingCase> describeDataRequestPendingCases(Integer requestId) throws NotFoundException;
-
-    List<DataOwnerCase> describeDataOwnerPendingCases(Integer dataOwnerId);
-}
