+++ src/test/java/org/broadinstitute/consent/http/resources/DataRequestCasesResourceTest.java
@@ -1,11 +1,9 @@
 package org.broadinstitute.consent.http.resources;
 
-import org.broadinstitute.consent.http.enumeration.ElectionType;
 import org.broadinstitute.consent.http.models.Summary;
-import org.broadinstitute.consent.http.service.AbstractPendingCaseAPI;
 import org.broadinstitute.consent.http.service.AbstractSummaryAPI;
 import org.broadinstitute.consent.http.service.ElectionService;
-import org.broadinstitute.consent.http.service.PendingCaseAPI;
+import org.broadinstitute.consent.http.service.PendingCaseService;
 import org.broadinstitute.consent.http.service.SummaryAPI;
 import org.junit.Assert;
 import org.junit.Before;
@@ -17,27 +15,25 @@ import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
-import javax.ws.rs.NotFoundException;
 import javax.ws.rs.ServerErrorException;
 import javax.ws.rs.core.Response;
 import java.util.Collections;
 import java.util.List;
 
 import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.anyInt;
+import static org.mockito.Matchers.anyObject;
 import static org.mockito.Mockito.when;
 
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({
-        AbstractPendingCaseAPI.class,
-        AbstractSummaryAPI.class
-})
+@PrepareForTest({AbstractSummaryAPI.class})
 public class DataRequestCasesResourceTest {
 
     @Mock
     ElectionService electionService;
 
     @Mock
-    PendingCaseAPI pendingCaseApi;
+    PendingCaseService pendingCaseService;
 
     @Mock
     SummaryAPI summaryApi;
@@ -47,15 +43,13 @@ public class DataRequestCasesResourceTest {
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
-        PowerMockito.mockStatic(AbstractPendingCaseAPI.class);
         PowerMockito.mockStatic(AbstractSummaryAPI.class);
-        when(AbstractPendingCaseAPI.getInstance()).thenReturn(pendingCaseApi);
         when(AbstractSummaryAPI.getInstance()).thenReturn(summaryApi);
     }
 
     @Test
     public void testGetDataRequestPendingCases() {
-        when(pendingCaseApi.describeDataRequestPendingCases(any())).thenReturn(Collections.emptyList());
+        when(pendingCaseService.describeDataRequestPendingCases(any())).thenReturn(Collections.emptyList());
         initResource();
         Response response = resource.getDataRequestPendingCases(null, null);
         Assert.assertEquals(200, response.getStatus());
@@ -65,7 +59,7 @@ public class DataRequestCasesResourceTest {
 
     @Test
     public void testGetDataOwnerPendingCases() {
-        when(pendingCaseApi.describeDataOwnerPendingCases(any())).thenReturn(Collections.emptyList());
+        when(pendingCaseService.describeDataOwnerPendingCases(anyInt(), anyObject())).thenReturn(Collections.emptyList());
         initResource();
         Response response = resource.getDataOwnerPendingCases(null, null);
         Assert.assertEquals(200, response.getStatus());
@@ -75,7 +69,7 @@ public class DataRequestCasesResourceTest {
 
     @Test
     public void testGetDataOwnerPendingCasesError() {
-        when(pendingCaseApi.describeDataOwnerPendingCases(any())).thenThrow(new ServerErrorException(500));
+        when(pendingCaseService.describeDataOwnerPendingCases(anyInt(), anyObject())).thenThrow(new ServerErrorException(500));
         initResource();
         Response response = resource.getDataOwnerPendingCases(null, null);
         Assert.assertEquals(500, response.getStatus());
@@ -112,7 +106,7 @@ public class DataRequestCasesResourceTest {
     }
 
     private void initResource() {
-        resource = new DataRequestCasesResource(electionService);
+        resource = new DataRequestCasesResource(electionService, pendingCaseService);
     }
 
 }
